// src/models/kit/raw-kit-asset.ts
/**
 * Copyright 2022 Gravwell, Inc. All rights reserved.
 *
 * Contact: [legal@gravwell.io](mailto:legal@gravwell.io)
 *
 * This software may be modified and distributed under the terms of the MIT
 * license. See the LICENSE file for details.
 */

export type RawKitAsset = {
	Type: 'image' | 'readme';
	Source: string;
	Legend: string;
	Featured: boolean;
	Banner: boolean;
};